requirements.txt
redis>=5.0.0
flask>=2.0.0
flask_sqlalchemy>=2.5.1
flask_marshmallow>=0.14.0
marshmallow_sqlalchemy>=0.26.1
psycopg2-binary>=2.8.6
flask_migrate>=3.1.0